Create Timetabling 15
This is the GAMS code for the timetabling problem with a headway of 15 minutes. The table of the transfer passengers does not show all transfers. If there were the same transfer between two lines more than once, GAMS marked it as en error. To execute this code, these transfers had to be summed up.