Add files via upload
This is the excel file i used to rename the 11 lines into the new 22 lines and to calculate the travel time between any two cities. Not for any two cities i calculated the travel time with a formula. Additionally i used this file to identify the feasible direct trips between the cities and therefore where passengers have to change lines.